Move ImmutableSortedSet types out
