+++ src/System.Collections.Immutable/src/System/Collections/Immutable/ImmutableSortedSet_1.cs
@@ -1152,233 +1152,6 @@ namespace System.Collections.Immutable
         }
 
         /// <summary>
-        /// Enumerates the contents of a binary tree.
-        /// </summary>
-        /// <remarks>
-        /// This struct can and should be kept in exact sync with the other binary tree enumerators: 
-        /// <see cref="ImmutableList{T}.Enumerator"/>, <see cref="ImmutableSortedDictionary{TKey, TValue}.Enumerator"/>, and <see cref="ImmutableSortedSet{T}.Enumerator"/>.
-        /// 
-        /// CAUTION: when this enumerator is actually used as a valuetype (not boxed) do NOT copy it by assigning to a second variable 
-        /// or by passing it to another method.  When this enumerator is disposed of it returns a mutable reference type stack to a resource pool,
-        /// and if the value type enumerator is copied (which can easily happen unintentionally if you pass the value around) there is a risk
-        /// that a stack that has already been returned to the resource pool may still be in use by one of the enumerator copies, leading to data
-        /// corruption and/or exceptions.
-        /// </remarks>
-        [EditorBrowsable(EditorBrowsableState.Advanced)]
-        public struct Enumerator : IEnumerator<T>, ISecurePooledObjectUser, IStrongEnumerator<T>
-        {
-            /// <summary>
-            /// The resource pool of reusable mutable stacks for purposes of enumeration.
-            /// </summary>
-            /// <remarks>
-            /// We utilize this resource pool to make "allocation free" enumeration achievable.
-            /// </remarks>
-            private static readonly SecureObjectPool<Stack<RefAsValueType<Node>>, Enumerator> s_enumeratingStacks =
-                new SecureObjectPool<Stack<RefAsValueType<Node>>, Enumerator>();
-
-            /// <summary>
-            /// The builder being enumerated, if applicable.
-            /// </summary>
-            private readonly Builder _builder;
-
-            /// <summary>
-            /// A unique ID for this instance of this enumerator.
-            /// Used to protect pooled objects from use after they are recycled.
-            /// </summary>
-            private readonly int _poolUserId;
-
-            /// <summary>
-            /// A flag indicating whether this enumerator works in reverse sort order.
-            /// </summary>
-            private readonly bool _reverse;
-
-            /// <summary>
-            /// The set being enumerated.
-            /// </summary>
-            private Node _root;
-
-            /// <summary>
-            /// The stack to use for enumerating the binary tree.
-            /// </summary>
-            /// <remarks>
-            /// We use <see cref="RefAsValueType{T}"/> as a wrapper to avoid paying the cost of covariant checks whenever
-            /// the underlying array that the <see cref="Stack{T}"/> class uses is written to. 
-            /// We've recognized this as a perf win in ETL traces for these stack frames:
-            /// clr!JIT_Stelem_Ref
-            ///   clr!ArrayStoreCheck
-            ///     clr!ObjIsInstanceOf
-            /// </remarks>
-            private SecurePooledObject<Stack<RefAsValueType<Node>>> _stack;
-
-            /// <summary>
-            /// The node currently selected.
-            /// </summary>
-            private Node _current;
-
-            /// <summary>
-            /// The version of the builder (when applicable) that is being enumerated.
-            /// </summary>
-            private int _enumeratingBuilderVersion;
-
-            /// <summary>
-            /// Initializes an <see cref="Enumerator"/> structure.
-            /// </summary>
-            /// <param name="root">The root of the set to be enumerated.</param>
-            /// <param name="builder">The builder, if applicable.</param>
-            /// <param name="reverse"><c>true</c> to enumerate the collection in reverse.</param>
-            internal Enumerator(Node root, Builder builder = null, bool reverse = false)
-            {
-                Requires.NotNull(root, nameof(root));
-
-                _root = root;
-                _builder = builder;
-                _current = null;
-                _reverse = reverse;
-                _enumeratingBuilderVersion = builder != null ? builder.Version : -1;
-                _poolUserId = SecureObjectPool.NewId();
-                _stack = null;
-                if (!s_enumeratingStacks.TryTake(this, out _stack))
-                {
-                    _stack = s_enumeratingStacks.PrepNew(this, new Stack<RefAsValueType<Node>>(root.Height));
-                }
-
-                this.PushNext(_root);
-            }
-
-            /// <inheritdoc/>
-            int ISecurePooledObjectUser.PoolUserId
-            {
-                get { return _poolUserId; }
-            }
-
-            /// <summary>
-            /// The current element.
-            /// </summary>
-            public T Current
-            {
-                get
-                {
-                    this.ThrowIfDisposed();
-                    if (_current != null)
-                    {
-                        return _current.Value;
-                    }
-
-                    throw new InvalidOperationException();
-                }
-            }
-
-            /// <summary>
-            /// The current element.
-            /// </summary>
-            object System.Collections.IEnumerator.Current
-            {
-                get { return this.Current; }
-            }
-
-            /// <summary>
-            /// Disposes of this enumerator and returns the stack reference to the resource pool.
-            /// </summary>
-            public void Dispose()
-            {
-                _root = null;
-                _current = null;
-                Stack<RefAsValueType<Node>> stack;
-                if (_stack != null && _stack.TryUse(ref this, out stack))
-                {
-                    stack.ClearFastWhenEmpty();
-                    s_enumeratingStacks.TryAdd(this, _stack);
-                    _stack = null;
-                }
-            }
-
-            /// <summary>
-            /// Advances enumeration to the next element.
-            /// </summary>
-            /// <returns>A value indicating whether there is another element in the enumeration.</returns>
-            public bool MoveNext()
-            {
-                this.ThrowIfDisposed();
-                this.ThrowIfChanged();
-
-                var stack = _stack.Use(ref this);
-                if (stack.Count > 0)
-                {
-                    Node n = stack.Pop().Value;
-                    _current = n;
-                    this.PushNext(_reverse ? n.Left : n.Right);
-                    return true;
-                }
-                else
-                {
-                    _current = null;
-                    return false;
-                }
-            }
-
-            /// <summary>
-            /// Restarts enumeration.
-            /// </summary>
-            public void Reset()
-            {
-                this.ThrowIfDisposed();
-
-                _enumeratingBuilderVersion = _builder != null ? _builder.Version : -1;
-                _current = null;
-                var stack = _stack.Use(ref this);
-                stack.ClearFastWhenEmpty();
-                this.PushNext(_root);
-            }
-
-            /// <summary>
-            /// Throws an <see cref="ObjectDisposedException"/> if this enumerator has been disposed.
-            /// </summary>
-            private void ThrowIfDisposed()
-            {
-                Contract.Ensures(_root != null);
-                Contract.EnsuresOnThrow<ObjectDisposedException>(_root == null);
-
-                // Since this is a struct, copies might not have been marked as disposed.
-                // But the stack we share across those copies would know.
-                // This trick only works when we have a non-null stack.
-                // For enumerators of empty collections, there isn't any natural
-                // way to know when a copy of the struct has been disposed of.
-
-                if (_root == null || (_stack != null && !_stack.IsOwned(ref this)))
-                {
-                    Requires.FailObjectDisposed(this);
-                }
-            }
-
-            /// <summary>
-            /// Throws an exception if the underlying builder's contents have been changed since enumeration started.
-            /// </summary>
-            /// <exception cref="System.InvalidOperationException">Thrown if the collection has changed.</exception>
-            private void ThrowIfChanged()
-            {
-                if (_builder != null && _builder.Version != _enumeratingBuilderVersion)
-                {
-                    throw new InvalidOperationException(SR.CollectionModifiedDuringEnumeration);
-                }
-            }
-
-            /// <summary>
-            /// Pushes this node and all its Left (or Right, if reversed) descendants onto the stack.
-            /// </summary>
-            /// <param name="node">The starting node to push onto the stack.</param>
-            private void PushNext(Node node)
-            {
-                Requires.NotNull(node, nameof(node));
-                var stack = _stack.Use(ref this);
-                while (!node.IsEmpty)
-                {
-                    stack.Push(new RefAsValueType<Node>(node));
-                    node = _reverse ? node.Right : node.Left;
-                }
-            }
-        }
-
-        /// <summary>
         /// An reverse enumerable of a sorted set.
         /// </summary>
         private class ReverseEnumerable : IEnumerable<T>
@@ -1418,828 +1191,6 @@ namespace System.Collections.Immutable
             IEnumerator IEnumerable.GetEnumerator()
             {
                 return this.GetEnumerator();
-            }
-        }
-
-        /// <summary>
-        /// A node in the AVL tree storing this set.
-        /// </summary>
-        [DebuggerDisplay("{_key}")]
-        internal sealed class Node : IBinaryTree<T>, IEnumerable<T>
-        {
-            /// <summary>
-            /// The default empty node.
-            /// </summary>
-            internal static readonly Node EmptyNode = new Node();
-
-            /// <summary>
-            /// The key associated with this node.
-            /// </summary>
-            private readonly T _key;
-
-            /// <summary>
-            /// A value indicating whether this node has been frozen (made immutable).
-            /// </summary>
-            /// <remarks>
-            /// Nodes must be frozen before ever being observed by a wrapping collection type
-            /// to protect collections from further mutations.
-            /// </remarks>
-            private bool _frozen;
-
-            /// <summary>
-            /// The depth of the tree beneath this node.
-            /// </summary>
-            private byte _height; // AVL tree max height <= ~1.44 * log2(maxNodes + 2)
-
-            /// <summary>
-            /// The number of elements contained by this subtree starting at this node.
-            /// </summary>
-            /// <remarks>
-            /// If this node would benefit from saving 4 bytes, we could have only a few nodes 
-            /// scattered throughout the graph actually record the count of nodes beneath them.
-            /// Those without the count could query their descendants, which would often short-circuit
-            /// when they hit a node that *does* include a count field.
-            /// </remarks>
-            private int _count;
-
-            /// <summary>
-            /// The left tree.
-            /// </summary>
-            private Node _left;
-
-            /// <summary>
-            /// The right tree.
-            /// </summary>
-            private Node _right;
-
-            /// <summary>
-            /// Initializes a new instance of the <see cref="ImmutableSortedSet{T}.Node"/> class
-            /// that is pre-frozen.
-            /// </summary>
-            private Node()
-            {
-                Contract.Ensures(this.IsEmpty);
-                _frozen = true; // the empty node is *always* frozen.
-            }
-
-            /// <summary>
-            /// Initializes a new instance of the <see cref="ImmutableSortedSet{T}.Node"/> class
-            /// that is not yet frozen.
-            /// </summary>
-            /// <param name="key">The value stored by this node.</param>
-            /// <param name="left">The left branch.</param>
-            /// <param name="right">The right branch.</param>
-            /// <param name="frozen">Whether this node is prefrozen.</param>
-            private Node(T key, Node left, Node right, bool frozen = false)
-            {
-                Requires.NotNullAllowStructs(key, nameof(key));
-                Requires.NotNull(left, nameof(left));
-                Requires.NotNull(right, nameof(right));
-                Debug.Assert(!frozen || (left._frozen && right._frozen));
-
-                _key = key;
-                _left = left;
-                _right = right;
-                _height = checked((byte)(1 + Math.Max(left._height, right._height)));
-                _count = 1 + left._count + right._count;
-                _frozen = frozen;
-            }
-
-            /// <summary>
-            /// Gets a value indicating whether this instance is empty.
-            /// </summary>
-            /// <value>
-            ///   <c>true</c> if this instance is empty; otherwise, <c>false</c>.
-            /// </value>
-            public bool IsEmpty
-            {
-                get { return _left == null; }
-            }
-
-            /// <summary>
-            /// Gets the height of the tree beneath this node.
-            /// </summary>
-            public int Height
-            {
-                get { return _height; }
-            }
-
-            /// <summary>
-            /// Gets the left branch of this node.
-            /// </summary>
-            public Node Left
-            {
-                get { return _left; }
-            }
-
-            /// <summary>
-            /// Gets the left branch of this node.
-            /// </summary>
-            IBinaryTree IBinaryTree.Left
-            {
-                get { return _left; }
-            }
-
-            /// <summary>
-            /// Gets the right branch of this node.
-            /// </summary>
-            public Node Right
-            {
-                get { return _right; }
-            }
-
-            /// <summary>
-            /// Gets the right branch of this node.
-            /// </summary>
-            IBinaryTree IBinaryTree.Right
-            {
-                get { return _right; }
-            }
-
-            /// <summary>
-            /// Gets the left branch of this node.
-            /// </summary>
-            IBinaryTree<T> IBinaryTree<T>.Left
-            {
-                get { return _left; }
-            }
-
-            /// <summary>
-            /// Gets the right branch of this node.
-            /// </summary>
-            IBinaryTree<T> IBinaryTree<T>.Right
-            {
-                get { return _right; }
-            }
-
-            /// <summary>
-            /// Gets the value represented by the current node.
-            /// </summary>
-            public T Value { get { return _key; } }
-
-            /// <summary>
-            /// Gets the number of elements contained by this subtree starting at this node.
-            /// </summary>
-            public int Count
-            {
-                get { return _count; }
-            }
-
-            /// <summary>
-            /// Gets the key.
-            /// </summary>
-            internal T Key
-            {
-                get { return _key; }
-            }
-
-            /// <summary>
-            /// Gets the maximum value in the collection, as defined by the comparer.
-            /// </summary>
-            /// <value>The maximum value in the set.</value>
-            internal T Max
-            {
-                get
-                {
-                    if (this.IsEmpty)
-                    {
-                        return default(T);
-                    }
-
-                    Node n = this;
-                    while (!n._right.IsEmpty)
-                    {
-                        n = n._right;
-                    }
-
-                    return n._key;
-                }
-            }
-
-            /// <summary>
-            /// Gets the minimum value in the collection, as defined by the comparer.
-            /// </summary>
-            /// <value>The minimum value in the set.</value>
-            internal T Min
-            {
-                get
-                {
-                    if (this.IsEmpty)
-                    {
-                        return default(T);
-                    }
-
-                    Node n = this;
-                    while (!n._left.IsEmpty)
-                    {
-                        n = n._left;
-                    }
-
-                    return n._key;
-                }
-            }
-
-            /// <summary>
-            /// Gets the element of the set at the given index.
-            /// </summary>
-            /// <param name="index">The 0-based index of the element in the set to return.</param>
-            /// <returns>The element at the given position.</returns>
-            internal T this[int index]
-            {
-                get
-                {
-                    Requires.Range(index >= 0 && index < this.Count, nameof(index));
-
-                    if (index < _left._count)
-                    {
-                        return _left[index];
-                    }
-
-                    if (index > _left._count)
-                    {
-                        return _right[index - _left._count - 1];
-                    }
-
-                    return _key;
-                }
-            }
-
-            #region IEnumerable<T> Members
-
-            /// <summary>
-            /// Returns an enumerator that iterates through the collection.
-            /// </summary>
-            /// <returns>
-            /// A <see cref="IEnumerator{T}"/> that can be used to iterate through the collection.
-            /// </returns>
-            public Enumerator GetEnumerator()
-            {
-                return new Enumerator(this);
-            }
-
-            /// <summary>
-            /// Returns an enumerator that iterates through the collection.
-            /// </summary>
-            /// <returns>
-            /// A <see cref="IEnumerator{T}"/> that can be used to iterate through the collection.
-            /// </returns>
-            [ExcludeFromCodeCoverage] // internal and never called, but here for the interface.
-            IEnumerator<T> IEnumerable<T>.GetEnumerator()
-            {
-                return this.GetEnumerator();
-            }
-
-            /// <summary>
-            /// Returns an enumerator that iterates through the collection.
-            /// </summary>
-            /// <returns>
-            /// A <see cref="IEnumerator{T}"/> that can be used to iterate through the collection.
-            /// </returns>
-            [ExcludeFromCodeCoverage] // internal and never called, but here for the interface.
-            IEnumerator IEnumerable.GetEnumerator()
-            {
-                return this.GetEnumerator();
-            }
-
-            #endregion
-
-            /// <summary>
-            /// Returns an enumerator that iterates through the collection.
-            /// </summary>
-            /// <param name="builder">The builder, if applicable.</param>
-            /// <returns>
-            /// A <see cref="IEnumerator{T}"/> that can be used to iterate through the collection.
-            /// </returns>
-            internal Enumerator GetEnumerator(Builder builder)
-            {
-                return new Enumerator(this, builder);
-            }
-
-            /// <summary>
-            /// See the <see cref="ICollection{T}"/> interface.
-            /// </summary>
-            internal void CopyTo(T[] array, int arrayIndex)
-            {
-                Requires.NotNull(array, nameof(array));
-                Requires.Range(arrayIndex >= 0, nameof(arrayIndex));
-                Requires.Range(array.Length >= arrayIndex + this.Count, nameof(arrayIndex));
-                foreach (var item in this)
-                {
-                    array[arrayIndex++] = item;
-                }
-            }
-
-            /// <summary>
-            /// See the <see cref="ICollection{T}"/> interface.
-            /// </summary>
-            internal void CopyTo(Array array, int arrayIndex)
-            {
-                Requires.NotNull(array, nameof(array));
-                Requires.Range(arrayIndex >= 0, nameof(arrayIndex));
-                Requires.Range(array.Length >= arrayIndex + this.Count, nameof(arrayIndex));
-
-                foreach (var item in this)
-                {
-                    array.SetValue(item, arrayIndex++);
-                }
-            }
-
-            /// <summary>
-            /// Adds the specified key to the tree.
-            /// </summary>
-            /// <param name="key">The key.</param>
-            /// <param name="comparer">The comparer.</param>
-            /// <param name="mutated">Receives a value indicating whether this node tree has mutated because of this operation.</param>
-            /// <returns>The new tree.</returns>
-            internal Node Add(T key, IComparer<T> comparer, out bool mutated)
-            {
-                Requires.NotNullAllowStructs(key, nameof(key));
-                Requires.NotNull(comparer, nameof(comparer));
-
-                if (this.IsEmpty)
-                {
-                    mutated = true;
-                    return new Node(key, this, this);
-                }
-                else
-                {
-                    Node result = this;
-                    int compareResult = comparer.Compare(key, _key);
-                    if (compareResult > 0)
-                    {
-                        var newRight = _right.Add(key, comparer, out mutated);
-                        if (mutated)
-                        {
-                            result = this.Mutate(right: newRight);
-                        }
-                    }
-                    else if (compareResult < 0)
-                    {
-                        var newLeft = _left.Add(key, comparer, out mutated);
-                        if (mutated)
-                        {
-                            result = this.Mutate(left: newLeft);
-                        }
-                    }
-                    else
-                    {
-                        mutated = false;
-                        return this;
-                    }
-
-                    return mutated ? MakeBalanced(result) : result;
-                }
-            }
-
-            /// <summary>
-            /// Removes the specified key from the tree.
-            /// </summary>
-            /// <param name="key">The key.</param>
-            /// <param name="comparer">The comparer.</param>
-            /// <param name="mutated">Receives a value indicating whether this node tree has mutated because of this operation.</param>
-            /// <returns>The new tree.</returns>
-            internal Node Remove(T key, IComparer<T> comparer, out bool mutated)
-            {
-                Requires.NotNullAllowStructs(key, nameof(key));
-                Requires.NotNull(comparer, nameof(comparer));
-
-                if (this.IsEmpty)
-                {
-                    mutated = false;
-                    return this;
-                }
-                else
-                {
-                    Node result = this;
-                    int compare = comparer.Compare(key, _key);
-                    if (compare == 0)
-                    {
-                        // We have a match.
-                        mutated = true;
-
-                        // If this is a leaf, just remove it 
-                        // by returning Empty.  If we have only one child,
-                        // replace the node with the child.
-                        if (_right.IsEmpty && _left.IsEmpty)
-                        {
-                            result = EmptyNode;
-                        }
-                        else if (_right.IsEmpty && !_left.IsEmpty)
-                        {
-                            result = _left;
-                        }
-                        else if (!_right.IsEmpty && _left.IsEmpty)
-                        {
-                            result = _right;
-                        }
-                        else
-                        {
-                            // We have two children. Remove the next-highest node and replace
-                            // this node with it.
-                            var successor = _right;
-                            while (!successor._left.IsEmpty)
-                            {
-                                successor = successor._left;
-                            }
-
-                            bool dummyMutated;
-                            var newRight = _right.Remove(successor._key, comparer, out dummyMutated);
-                            result = successor.Mutate(left: _left, right: newRight);
-                        }
-                    }
-                    else if (compare < 0)
-                    {
-                        var newLeft = _left.Remove(key, comparer, out mutated);
-                        if (mutated)
-                        {
-                            result = this.Mutate(left: newLeft);
-                        }
-                    }
-                    else
-                    {
-                        var newRight = _right.Remove(key, comparer, out mutated);
-                        if (mutated)
-                        {
-                            result = this.Mutate(right: newRight);
-                        }
-                    }
-
-                    return result.IsEmpty ? result : MakeBalanced(result);
-                }
-            }
-
-            /// <summary>
-            /// Determines whether the specified key is in this tree.
-            /// </summary>
-            /// <param name="key">The key.</param>
-            /// <param name="comparer">The comparer.</param>
-            /// <returns>
-            ///   <c>true</c> if the tree contains the specified key; otherwise, <c>false</c>.
-            /// </returns>
-            [Pure]
-            internal bool Contains(T key, IComparer<T> comparer)
-            {
-                Requires.NotNullAllowStructs(key, nameof(key));
-                Requires.NotNull(comparer, nameof(comparer));
-                return !this.Search(key, comparer).IsEmpty;
-            }
-
-            /// <summary>
-            /// Freezes this node and all descendant nodes so that any mutations require a new instance of the nodes.
-            /// </summary>
-            internal void Freeze()
-            {
-                // If this node is frozen, all its descendants must already be frozen.
-                if (!_frozen)
-                {
-                    _left.Freeze();
-                    _right.Freeze();
-                    _frozen = true;
-                }
-            }
-
-            /// <summary>
-            /// Searches for the specified key.
-            /// </summary>
-            /// <param name="key">The key to search for.</param>
-            /// <param name="comparer">The comparer.</param>
-            /// <returns>The matching node, or <see cref="EmptyNode"/> if no match was found.</returns>
-            [Pure]
-            internal Node Search(T key, IComparer<T> comparer)
-            {
-                Requires.NotNullAllowStructs(key, nameof(key));
-                Requires.NotNull(comparer, nameof(comparer));
-
-                if (this.IsEmpty)
-                {
-                    return this;
-                }
-                else
-                {
-                    int compare = comparer.Compare(key, _key);
-                    if (compare == 0)
-                    {
-                        return this;
-                    }
-                    else if (compare > 0)
-                    {
-                        return _right.Search(key, comparer);
-                    }
-                    else
-                    {
-                        return _left.Search(key, comparer);
-                    }
-                }
-            }
-
-            /// <summary>
-            /// Searches for the specified key.
-            /// </summary>
-            /// <param name="key">The key to search for.</param>
-            /// <param name="comparer">The comparer.</param>
-            /// <returns>The matching node, or <see cref="EmptyNode"/> if no match was found.</returns>
-            [Pure]
-            internal int IndexOf(T key, IComparer<T> comparer)
-            {
-                Requires.NotNullAllowStructs(key, nameof(key));
-                Requires.NotNull(comparer, nameof(comparer));
-
-                if (this.IsEmpty)
-                {
-                    return -1;
-                }
-                else
-                {
-                    int compare = comparer.Compare(key, _key);
-                    if (compare == 0)
-                    {
-                        return _left.Count;
-                    }
-                    else if (compare > 0)
-                    {
-                        int result = _right.IndexOf(key, comparer);
-                        bool missing = result < 0;
-                        if (missing)
-                        {
-                            result = ~result;
-                        }
-
-                        result = _left.Count + 1 + result;
-                        if (missing)
-                        {
-                            result = ~result;
-                        }
-
-                        return result;
-                    }
-                    else
-                    {
-                        return _left.IndexOf(key, comparer);
-                    }
-                }
-            }
-
-            /// <summary>
-            /// Returns an <see cref="IEnumerable{T}"/> that iterates over this
-            /// collection in reverse order.
-            /// </summary>
-            /// <returns>
-            /// An enumerator that iterates over the <see cref="ImmutableSortedSet{T}"/>
-            /// in reverse order.
-            /// </returns>
-            [Pure]
-            internal IEnumerator<T> Reverse()
-            {
-                return new Enumerator(this, reverse: true);
-            }
-
-            #region Tree balancing methods
-
-            /// <summary>
-            /// AVL rotate left operation.
-            /// </summary>
-            /// <param name="tree">The tree.</param>
-            /// <returns>The rotated tree.</returns>
-            private static Node RotateLeft(Node tree)
-            {
-                Requires.NotNull(tree, nameof(tree));
-                Debug.Assert(!tree.IsEmpty);
-                Contract.Ensures(Contract.Result<Node>() != null);
-
-                if (tree._right.IsEmpty)
-                {
-                    return tree;
-                }
-
-                var right = tree._right;
-                return right.Mutate(left: tree.Mutate(right: right._left));
-            }
-
-            /// <summary>
-            /// AVL rotate right operation.
-            /// </summary>
-            /// <param name="tree">The tree.</param>
-            /// <returns>The rotated tree.</returns>
-            private static Node RotateRight(Node tree)
-            {
-                Requires.NotNull(tree, nameof(tree));
-                Debug.Assert(!tree.IsEmpty);
-                Contract.Ensures(Contract.Result<Node>() != null);
-
-                if (tree._left.IsEmpty)
-                {
-                    return tree;
-                }
-
-                var left = tree._left;
-                return left.Mutate(right: tree.Mutate(left: left._right));
-            }
-
-            /// <summary>
-            /// AVL rotate double-left operation.
-            /// </summary>
-            /// <param name="tree">The tree.</param>
-            /// <returns>The rotated tree.</returns>
-            private static Node DoubleLeft(Node tree)
-            {
-                Requires.NotNull(tree, nameof(tree));
-                Debug.Assert(!tree.IsEmpty);
-                Contract.Ensures(Contract.Result<Node>() != null);
-
-                if (tree._right.IsEmpty)
-                {
-                    return tree;
-                }
-
-                Node rotatedRightChild = tree.Mutate(right: RotateRight(tree._right));
-                return RotateLeft(rotatedRightChild);
-            }
-
-            /// <summary>
-            /// AVL rotate double-right operation.
-            /// </summary>
-            /// <param name="tree">The tree.</param>
-            /// <returns>The rotated tree.</returns>
-            private static Node DoubleRight(Node tree)
-            {
-                Requires.NotNull(tree, nameof(tree));
-                Debug.Assert(!tree.IsEmpty);
-                Contract.Ensures(Contract.Result<Node>() != null);
-
-                if (tree._left.IsEmpty)
-                {
-                    return tree;
-                }
-
-                Node rotatedLeftChild = tree.Mutate(left: RotateLeft(tree._left));
-                return RotateRight(rotatedLeftChild);
-            }
-
-            /// <summary>
-            /// Returns a value indicating whether the tree is in balance.
-            /// </summary>
-            /// <param name="tree">The tree.</param>
-            /// <returns>0 if the tree is in balance, a positive integer if the right side is heavy, or a negative integer if the left side is heavy.</returns>
-            [Pure]
-            private static int Balance(Node tree)
-            {
-                Requires.NotNull(tree, nameof(tree));
-                Debug.Assert(!tree.IsEmpty);
-
-                return tree._right._height - tree._left._height;
-            }
-
-            /// <summary>
-            /// Determines whether the specified tree is right heavy.
-            /// </summary>
-            /// <param name="tree">The tree.</param>
-            /// <returns>
-            /// <c>true</c> if [is right heavy] [the specified tree]; otherwise, <c>false</c>.
-            /// </returns>
-            [Pure]
-            private static bool IsRightHeavy(Node tree)
-            {
-                Requires.NotNull(tree, nameof(tree));
-                Debug.Assert(!tree.IsEmpty);
-                return Balance(tree) >= 2;
-            }
-
-            /// <summary>
-            /// Determines whether the specified tree is left heavy.
-            /// </summary>
-            [Pure]
-            private static bool IsLeftHeavy(Node tree)
-            {
-                Requires.NotNull(tree, nameof(tree));
-                Debug.Assert(!tree.IsEmpty);
-                return Balance(tree) <= -2;
-            }
-
-            /// <summary>
-            /// Balances the specified tree.
-            /// </summary>
-            /// <param name="tree">The tree.</param>
-            /// <returns>A balanced tree.</returns>
-            [Pure]
-            private static Node MakeBalanced(Node tree)
-            {
-                Requires.NotNull(tree, nameof(tree));
-                Debug.Assert(!tree.IsEmpty);
-                Contract.Ensures(Contract.Result<Node>() != null);
-
-                if (IsRightHeavy(tree))
-                {
-                    return Balance(tree._right) < 0 ? DoubleLeft(tree) : RotateLeft(tree);
-                }
-
-                if (IsLeftHeavy(tree))
-                {
-                    return Balance(tree._left) > 0 ? DoubleRight(tree) : RotateRight(tree);
-                }
-
-                return tree;
-            }
-
-            #endregion
-
-            /// <summary>
-            /// Creates a node tree that contains the contents of a list.
-            /// </summary>
-            /// <param name="items">An indexable list with the contents that the new node tree should contain.</param>
-            /// <param name="start">The starting index within <paramref name="items"/> that should be captured by the node tree.</param>
-            /// <param name="length">The number of elements from <paramref name="items"/> that should be captured by the node tree.</param>
-            /// <returns>The root of the created node tree.</returns>
-            [Pure]
-            internal static Node NodeTreeFromList(IOrderedCollection<T> items, int start, int length)
-            {
-                Requires.NotNull(items, nameof(items));
-                Debug.Assert(start >= 0);
-                Debug.Assert(length >= 0);
-
-                if (length == 0)
-                {
-                    return EmptyNode;
-                }
-
-                int rightCount = (length - 1) / 2;
-                int leftCount = (length - 1) - rightCount;
-                Node left = NodeTreeFromList(items, start, leftCount);
-                Node right = NodeTreeFromList(items, start + leftCount + 1, rightCount);
-                return new Node(items[start + leftCount], left, right, true);
-            }
-
-            /// <summary>
-            /// Creates a node mutation, either by mutating this node (if not yet frozen) or by creating a clone of this node
-            /// with the described changes.
-            /// </summary>
-            /// <param name="left">The left branch of the mutated node.</param>
-            /// <param name="right">The right branch of the mutated node.</param>
-            /// <returns>The mutated (or created) node.</returns>
-            private Node Mutate(Node left = null, Node right = null)
-            {
-                if (_frozen)
-                {
-                    return new Node(_key, left ?? _left, right ?? _right);
-                }
-                else
-                {
-                    if (left != null)
-                    {
-                        _left = left;
-                    }
-
-                    if (right != null)
-                    {
-                        _right = right;
-                    }
-
-                    _height = checked((byte)(1 + Math.Max(_left._height, _right._height)));
-                    _count = 1 + _left._count + _right._count;
-                    return this;
-                }
-            }
-        }
-    }
-
-    /// <summary>
-    /// A simple view of the immutable collection that the debugger can show to the developer.
-    /// </summary>
-    internal class ImmutableSortedSetDebuggerProxy<T>
-    {
-        /// <summary>
-        /// The collection to be enumerated.
-        /// </summary>
-        private readonly ImmutableSortedSet<T> _set;
-
-        /// <summary>
-        /// The simple view of the collection.
-        /// </summary>
-        private T[] _contents;
-
-        /// <summary>   
-        /// Initializes a new instance of the <see cref="ImmutableSortedSetDebuggerProxy{T}"/> class.
-        /// </summary>
-        /// <param name="set">The collection to display in the debugger</param>
-        public ImmutableSortedSetDebuggerProxy(ImmutableSortedSet<T> set)
-        {
-            Requires.NotNull(set, nameof(set));
-            _set = set;
-        }
-
-        /// <summary>
-        /// Gets a simple debugger-viewable collection.
-        /// </summary>
-        [DebuggerBrowsable(DebuggerBrowsableState.RootHidden)]
-        public T[] Contents
-        {
-            get
-            {
-                if (_contents == null)
-                {
-                    _contents = _set.ToArray(_set.Count);
-                }
-
-                return _contents;
             }
         }
     }
